// src/app/models/scheme_details.model.ts
export interface ISchemeDetails {
  productname: string;
  tenure: string;
  terms: string;
  termsFrequancy: string;
  maturityDate: string;
  baseDepositamount: string;
  instalmentAmount: string;
  instalmentFrequancy: string;
  maturityAmount: string;
  interestRate: string;
  bonusAmount: string;
}