// src/app/models/time_info.model.ts
export interface ITimeInfo {
  principalAmount: string;
  principalAmountMaxValue: number;
  principalAmountMinValue: number;
  principalAmountReadonly: boolean;
  principalAmountToolTip: string;
  termMaxValue: number;
  termMinvalue: number;
  termReadonly: boolean;
  termToolTip: string;
  terms: string;
}