// src/app/models/bank_homebankbranchinfo.model.ts
export interface IBranchInfo {
    value: string;
}

export interface IHomeBranchInfoResponse {
    maturitY_DT?: null;
    sanctioN_DT?: null;
    disbursemenT_DT?: null;
    displaY_NAME_CONV?: null;
    rownum?: null;
    producT_ID?: null;
    producT_NM?: null;
    brancH_ID?: string;
    basE_CURRENCY_ID?: string;
    brancH_NM?: string;
    islamiC_FLAG?: string;
    offshorE_FLAG?: string;
    accounT_NUMBER?: null;
    accounT_TITLE?: null;
    accounT_STATUS?: null;
    currencY_NM?: null;
    availablE_BALANCE?: null;
    opeN_DT?: null;
    lasT_TRN_DT?: null;
}